court_data_fetcher/requirements.txt
Flask==2.3.3
selenium==4.15.2
beautifulsoup4==4.12.2
mysql-connector-python==8.4.0
